requirements.txt
Django==1.10.2
django-jet==1.0.2
django-tinymce
